% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndvi.R
\name{ndvi}
\alias{ndvi}
\title{Normalization Difference Vegetation Index}
\usage{
ndvi(nclass, opossite = FALSE, img_prev = FALSE)
}
\arguments{
\item{nclass}{number of break of palette.}

\item{opossite}{change the position of the palette.}

\item{img_prev}{previous visualization of palette.}
}
\value{
Hexadecimal code of colour
}
\description{
Mapping the ndvi index
}
\examples{
\dontrun{
ndvi_index <- ndvi(nclass = 5, opossite = TRUE, img_prev = TRUE)
}
}
